rpcn: pretty print linter errors on update
